Create Chijioke-chinazaekpere-2024714114
This python program implements a grading system using 'if' and 'elif' statements. It takes a students score as input and prints the corresponding grades based on the following scale: 


- 70- 100: GradeA
- 60- 69: Grade B
- 50-59: Grade C
- 45-49: Grade D
- 40-44: Grade E
- 20- 39: Grade F

For each valid score, the program also prints "YES" to confirm the grading was successful.  If the score is outside. If the score is outside the range it prints "invalid score" or "No"

Created by: chijioke chinazaekpere 
Reg no: 2024714114